Factor out IDL module
